Implementing the visitor pattern for the parser
Due to the current design, we cannot have all the different expressions
as structs implementing the Expr trait. This would hinder implementing
the visitor pattern with the following error:

error[E0038]: the trait `Expr` is not dyn compatible
 --> src/parser.rs:9:15
  |
9 |     left: Box<dyn Expr>,
  |               ^^^^^^^^ `Expr` is not dyn compatible
  |
note: for a trait to be dyn compatible it needs to allow building a vtable
      for more information, visit <https://doc.rust-lang.org/reference/items/traits.html#dyn-compatibility>
 --> src/parser.rs:5:8
  |
4 | pub trait Expr {
  |           ---- this trait is not dyn compatible...
5 |     fn accept<T>(&self, visitor: &dyn Visitor<T>) -> T;
  |        ^^^^^^ ...because method `accept` has generic type parameters

This is a intentionally bogus commit for further reference on the issue.
Next, we should convert the Expr trait into an Enum.