Use enum to represent expressions
Moving away from structs to represent our expressions makes it simpler
to implement the visitor pattern suggested by the book. However, it is
clear that the indirections are not needed given we have access to
rust's pattern matching. Next, we should refactor the parser to remove
the indirections as suggested by the challenges in the end of Chapter 5.